mlxsw: spectrum_trap: Dynamically allocate trap groups array
A later patch will allow each ASIC to register its own trap groups, as
not all trap groups are supported by all ASICs. Each ASIC will have an
array of trap groups unique to it, in addition to the array of common
trap groups (i.e., 'mlxsw_sp_trap_group_items_arr').

Duplicate the array of common trap groups to the per-ASIC private struct
so that the unique trap groups could be copied to it in the per-ASIC
initialization.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>