mlxsw: spectrum_trap: Dynamically allocate traps array
A later patch will allow each ASIC to register its traps, as not all
traps are supported by all ASICs. Each ASIC will have an array of traps
unique to it, in addition to the array of common traps (i.e.,
'mlxsw_sp_trap_items_arr').

Duplicate the array of common traps to the per-ASIC private struct so
that the unique traps could be copied to it in the per-ASIC
initialization.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>