mlxsw: spectrum_trap: Warn if burst size does not fit requested rate
In order to avoid intermittent drops by the packet trap policers when
the incoming packet rate is slightly lower than the configured policer
rate, the minimum policer burst size needs to be limited according to
the configured rate.

For Spectrum-1 the minimum burst size is 5% of the rate.
For Spectrum-{2,3} the minimum burst size is 40% of the rate.

Emit a warning to user space via extack in case the burst size is too
low.

Fixes: 13f2e64b94ea ("mlxsw: spectrum_trap: Add devlink-trap policer support")
Signed-off-by: Ido Schimmel <idosch@mellanox.com>