devlink: Add a raw tracepoint for trap reports
Add a tracepoint for trap reports so that drop monitor could register
its probe on it. Use trace_devlink_trap_report_enabled() to avoid
wasting cycles setting the trap metadata if the tracepoint is not
enabled.

The tracepoint is added as a raw tracepoint so that it is not part of
kernel ABI.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>