ethtool: Fix incompatibility between netlink and ioctl interfaces
With the ioctl interface, when autoneg is enabled, but without
specifying speed, duplex or link modes, the advertised link modes are
set to the supported link modes by the ethtool user space utility.
Example:

# ethtool --version
# ethtool -s swp3 speed 100000 autoneg off
# ethtool swp3
# ethtool -s swp3 autoneg on
# ethtool swp3

With the netlink interface, the same thing is done by the kernel, but
only if speed or duplex are specified. In which case, the advertised
link modes are set by traversing the supported link modes and picking
the ones matching the specified speed or duplex.

However, if speed nor duplex are specified, the driver is passed an
empty advertised link modes bitmap. This causes the mlxsw driver to
return an error. Other drivers might also be affected. Example:

# ethtool --version
# ethtool -s swp3 speed 100000 autoneg off
# ethtool swp3
# ethtool -s swp3 autoneg on
# ethtool swp3

Fix this incompatibility problem by having the kernel set the advertised
link modes according to the supported link modes even if speed nor
duplex are specified. In which case, the advertised link modes are set
to the supported link modes. Example after this patch:

# ethtool --version
# ethtool -s swp3 speed 100000 autoneg off
# ethtool swp3
# ethtool -s swp3 autoneg on
# ethtool swp3

Fixes: bfbcfe2032e7 ("ethtool: set link modes related data with LINKMODES_SET request")
Signed-off-by: Ido Schimmel <idosch@nvidia.com>