mlxsw: reg: Extend MGPIR register with new slot number field
Extend MGPIR (Management General Peripheral Information Register) with
new field specifying the slot number. The purpose of this field is to
support access to MPGIR register on modular system for getting the
number of cage, equipped on the line card, inserted at specified slot.
In case slot number is set zero, MGPIR will provide the information for
the main board. For Top of the Rack (non-modular) system it will
provide the same as before.

Signed-off-by: Vadim Pasternak <vadimp@nvidia.com>
Reviewed-by: Jiri Pirko <jiri@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>