nexthop: Add netlink handlers for bucket get
The hash table associated with a resilient next hop group cannot be dumped
together with the group itself, because there might be too many bucket to
fit the information in one netlink message. Instead, allow getting (but not
setting) individual buckets to inspect the next hop mapped therein, idle
time, and flags.

Signed-off-by: Petr Machata <petrm@nvidia.com>