nexthop: Extract a helper for validation of get/del RTNL requests
Validation of messages for get / del of a netxthop is the same as will be
validation of messages for get of a resilient next hop group bucket. The
difference is that policy for resilient next hop group buckets is a
superset of that used for nexthop get.

It is therefore possible to reuse the code that validates the message as
such, extract the next hop ID, and validates that. To that end, extract
from nh_valid_get_del_req() a helper __nh_valid_get_del_req() that does
just that.

Signed-off-by: Petr Machata <petrm@nvidia.com>