nexthop: Assert the invariant that a NH group is of only one type
Most of the code that deals with nexthop groups relies on the fact that the
group is of exactly one well-known type. Currently there is only one flag,
"mpath", but as more NHG types come, more flags will pop up, and it is
desirable to have a central place where their setting is validated.
Introduce such place into nexthop_create_group(), so that the check is done
before the code that relies on that invariant is invoked.

Signed-off-by: Petr Machata <petrm@nvidia.com>