mlxsw: spectrum_trap: Use common listener for global sampling triggers
Currently, the only supported global sampling trigger (i.e., not local
to a specific port) is the policy engine trigger. It is triggered
whenever packets are sampled in the policy engine.

Subsequent patches are going to add more global sampling triggers.
Packets sampled from these triggers will need to be processed in a very
similar manner to packets sampled from the policy engine, so factor out
a common listener function that takes the sampling trigger type as one
of its arguments.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>