mlxsw: spectrum_trap: Prepare Rx listeners for buffer related sampling
Packets that are sampled by attaching filters with sample action to the
'early_drop' and 'mark' qevents need to be sent to the psample module by
calling psample_sample_packet().

Adjust the Rx listeners for the early drop and ECN marking mirroring
reasons to call into psample if there is an active sampling trigger for
the mirroring reason.

A new SPAN session identifier is added for such packets in order to
distinguish them from packets that are sent to the CPU due to trap
action. The new identifier ensures that trap and sample action with a
rate of 1 cannot be used simultaneously for the same qevent. Such a
configuration will be rejected by mlxsw_sp_span_agent_bind() as only a
single SPAN agent can be bound to a global trigger.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>