mlxsw: spectrum_qdisc: Prepare for mirroring with probability
Currently, when enabled, every ECN marked or early dropped packet is
mirrored towards the CPU. Follow-up patches are going to add support for
mirroring with probability via sample action.

Prepare for this change by adding probability rate argument to
mlxsw_sp_qevent_span_configure(). For now, it is always set to 1, which
means that every packet is mirrored.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>