mlxsw: spectrum_qdisc: Add support buffer related sampling
Allow sample action to be attached to qevents with matchall classifier.
This is very similar to trap action support with two main differences.

First, the sampling parameters (e.g., psample group, rate) are stored in
the sampling triggers hash table so that they could be retrieved upon
incoming packets.

Second, the sampling rate towards the CPU port is not blindly set to 1
like with the trap action. Instead, the user supplied rate is used.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>