mlxsw: spectrum_qdisc: Track permissible actions per binding
One block can be bound to several qevents. The qevent type that the block
is bound to determines which actions make sense in a given context.
trap_fwd for example cannot be offloaded on a RED drop qevent, because the
trap_fwd contract specifies that the packet is forwarded in the HW
datapath, and the HW trigger that the action is offloaded to is always
dropping the packet.

Therefore keep track of which actions are permissible at each binding
block. When an attempt is made to bind a certain action at a binding point
where it is not supported, bounce the request.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>