mlxsw: Track per-module port status
In the common port module core, track the number of logical ports that
are mapped to the port module and the number of logical ports using it
that are administratively up.

This will be used by later patches to potentially veto certain
operations on the module, such as reset and setting its power mode.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>