ethtool: Add transceiver module extended states
Add an extended state and two extended sub-states to describe link
issues related to transceiver modules.

The first, 'ETHTOOL_LINK_EXT_SUBSTATE_MODULE_LOW_POWER_MODE', tells user
space that port is unable to gain a carrier because the associated
transceiver module is in low power mode where the data path is
deactivated. This is applicable to both SFF-8636 and CMIS modules.

Example:

 # ethtool --set-module swp13 power-mode-policy low

 # ip link set dev swp13 up

 $ ethtool swp13
 ...
 Link detected: no (Module, Module is in low power mode)

 # ip link set dev swp13 down

 # ethtool --set-module swp13 power-mode-policy high-on-up

 # ip link set dev swp13 up

 $ ethtool swp13
 ...
 Link detected: yes

The second, 'ETHTOOL_LINK_EXT_SUBSTATE_MODULE_CMIS_NOT_READY', tells
user space that port is unable to gain a carrier because the CMIS Module
State Machine did not reach the ModuleReady (Fully Operational) state.
For example, if the module is stuck at ModuleFault state. In which case,
user space can read the fault reason from the module's EEPROM and
potentially reset it.

For CMIS modules, the first extended sub-state is contained in the
second, but has the added advantage of being applicable to more module
types and being more specific about the nature of the problem.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>