TMP: mvpp2: Add shutdown method in mvpp2 driver
On an arm64 platform with the Spectrum ASIC, after loading and executing
a new kernel via kexec, bugs that relates to mvpp2 driver were observed.
This seems to be caused by the fact that the device is not properly
shutdown before executing the new kernel.

Fix this by implementing a shutdown method which mirrors the remove
method, as recommended by the kexec maintainer [1][2].

[1]
https://www.mail-archive.com/linux-kernel@vger.kernel.org/msg1195432.html

[2]
https://patchwork.kernel.org/project/linux-scsi/patch/20170212214920.28866-1-anton@ozlabs.org/#20116693

Signed-off-by: Danielle Ratson <danieller@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>