mlxsw: Split code of handling tunnel entries per address family
Currently, the function which adds/removes tunnel unicast entry to FDB
is shared between IPv4 and IPv6, while for IPv6 it warns because there is
no support for it.

The code for IPv6 will be more complicated because there is a need to
allocate/release entry in KVDL for IPv6 address.

As preparation for IPv6 underlay support, split the code according to
address family and split SFD filling as well. For now, warn for IPv6 entry.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>