mlxsw: spectrum_nve: Add API for maintaining IPv6 address per {FID, MAC}
When FID supposed to be disabled, all the relevant FDB entries should be
flushed also. Currently, the flush is done using SFDF register which can
flush all entries per FID.

For IPv6 tunnel entries, the IPv6 address is saved as part of KVDL, so
there is a need to be able to free all the relevant IPv6 entries also.

Add hash table to save mapping between {FID,MAC}->{IPv6}, and save all
the nodes in a list, so when FID will be disabled, it will be possible
to iterate over the list and free all the relevant IPv6 addresses.

The MAC is used as part of the key to enable identification of entry
replacement, in such case, the previous IPv6 address needs to be
released and entry for the new one should be allocated.

Add API to add and delete IPv6 address mapping which will be used in the
next patch.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>